incorporating it to bionc core
