+++ bionc/bionc_numpy/natural_segment.py
@@ -804,3 +804,52 @@ class NaturalSegment(AbstractNaturalSegment):
             Kinetic energy of the segment
         """
         return 0.5 * transpose(Qdoti.to_array()) @ self.mass_matrix @ Qdoti.to_array()
+
+    def _one_segment_inverse_dynamics(
+            self,
+            Qi: SegmentNaturalCoordinates,
+            Qddoti: SegmentNaturalAccelerations,
+            subtree_intersegmental_generalized_forces: np.ndarray,
+            segment_external_forces: np.ndarray,
+    ):
+        """
+        This function computes the inverse dynamics of a segment.
+
+        Parameters
+        ----------
+        Qi: SegmentNaturalCoordinates
+            The generalized coordinates of the segment
+        Qddoti: SegmentNaturalAccelerations
+            The generalized accelerations of the segment
+        subtree_intersegmental_generalized_forces : np.ndarray
+            The generalized forces applied to the segment by its children
+        segment_external_forces : np.ndarray
+            The generalized forces applied to the segment by the external forces
+
+        Returns
+        -------
+        force: np.ndarray
+            The force generated by the segment
+        torque: np.ndarray
+            The torque generated by the segment
+        lambdas: np.ndarray
+            The forces generated by the rigid body constraints
+        """
+
+        proximal_interpolation_matrix = NaturalVector.proximal().interpolate()
+        pseudo_interpolation_matrix = Qi.compute_pseudo_interpolation_matrix()
+        rigid_body_constraints_jacobian = self.rigid_body_constraint_jacobian(Qi=Qi)
+
+        # make a matrix out of it, todo: would be great to know if there is an analytical way to compute this matrix
+        front_matrix = np.hstack(
+            (proximal_interpolation_matrix.T, pseudo_interpolation_matrix.T, -rigid_body_constraints_jacobian.T))
+
+        # compute the generalized forces
+        generalized_forces = np.linalg.inv(front_matrix) @ (
+                (self.mass_matrix @ Qddoti)[:, np.newaxis]
+                - self.gravity_force()[:, np.newaxis]
+                - segment_external_forces
+                - subtree_intersegmental_generalized_forces
+        )
+
+        return generalized_forces[:3, 0], generalized_forces[3:6, 0], generalized_forces[6:, 0]
