+++ bionc/protocols/biomechanical_model.py
@@ -7,6 +7,8 @@ import dill as pickle
 
 from bionc.protocols.natural_coordinates import NaturalCoordinates
 from bionc.protocols.natural_velocities import NaturalVelocities
+from bionc.protocols.natural_accelerations import NaturalAccelerations
+from bionc.protocols.external_force import ExternalForceList
 
 
 class GenericBiomechanicalModel(ABC):
@@ -795,4 +797,64 @@ class GenericBiomechanicalModel(ABC):
         -------
             The position of the center of mass [3, nbSegments]
             in the global coordinate system/ inertial coordinate system
+        """
+
+    def _depth_first_search(self, segment_index, visited_segments=None) -> list[bool]:
+        """
+        This function returns the segments in a depth first search order.
+
+        todo: generalize to any number of segments with no parent.
+
+        Parameters
+        ----------
+        segment_index: int
+            The index of the segment to start the search from
+        visited_segments: list[Segment]
+            The segments already visited
+
+        Returns
+        -------
+        list[bool, ...
+            The segments in a depth first search order
+        """
+        if visited_segments is None:
+            visited_segments = [False for _ in range(self.nb_segments)]
+
+        visited_segments[segment_index] = True
+        for child_index in self.children(segment_index):
+            if visited_segments[child_index]:
+                raise RuntimeError("The model contain closed loops, we cannot use this algorithm")
+            if not visited_segments[child_index]:
+                visited_segments = self._depth_first_search(child_index, visited_segments)
+
+        return visited_segments
+
+    @abstractmethod
+    def inverse_dynamics(self, Q:NaturalCoordinates, Qddot: NaturalAccelerations,  external_forces: ExternalForceList = None):
+        """
+        This function returns the forces, torques and lambdas computes through recursive Newton-Euler algorithm
+
+        Source
+        ------
+        Dumas. R and Chèze. L (2006).
+        3D inverse dynamics in non-orthonormal segment coordinate system. Med Bio Eng Comput.
+        DOI 10.1007/s11517-006-0156-8
+
+        Parameters
+        ----------
+        Q: NaturalCoordinates
+           The generalized coordinates of the model
+        Qddot: NaturalAccelerations
+           The generalized accelerations of the model
+        external_forces: ExternalForceList
+           The external forces applied to the model
+
+        Returns
+        -------
+        torques: np.ndarray
+           The intersegmental torques
+        forces: np.ndarray
+           The intersegmental forces
+        lambdas: np.ndarray
+           The lagrange multipliers due to rigid contacts constraints
         """
